refactor: export all custom SVG icons from single file
