+++ site/src/components/icons/BlockProtocolIcon.tsx
