+++ site/src/components/icons/BlockProtocolLogoIcon.tsx
