+++ site/src/components/icons/BoltIcon.tsx
