+++ site/src/components/icons/CollapseIcon.tsx
