+++ site/src/components/icons/CopyIcon.tsx
