+++ site/src/components/icons/HASHLogoIcon.tsx
