+++ site/src/components/icons/SpecificationIcon.tsx
