+++ site/src/components/icons/SyncIcon.tsx
