+++ site/src/components/pages/dashboard/GenerateApiModal.tsx
@@ -2,7 +2,7 @@ import { Box, Typography } from "@mui/material";
 import { VoidFunctionComponent, ChangeEvent, FormEvent, useState } from "react";
 import { Button } from "../../Button";
 import { Modal } from "../../Modal";
-import { WarningIcon } from "../../SvgIcon/WarningIcon";
+import { WarningIcon } from "../../icons";
 import { ApiKeyRenderer } from "./ApiKeyRenderer";
 import { apiClient } from "../../../lib/apiClient";
